refactor: hooks/ 정리
