+++ src/todo/item/TodoItem.tsx
@@ -1,10 +1,10 @@
 import { Todo } from "../../routes/api"
 import { Modify } from "../containers/Modify"
-import { useBoolean } from "../hooks"
 import { ModifyButton, RemoveButton } from "./components/buttons"
 import { TodoCheckbox } from "./components/checkbox"
 import { Content } from "./components/views"
 import classes from "./TodoItem.module.css"
+import { useBoolean } from "./useBoolean"
 
 export const TodoItem = (item: Todo) => {
   const { val: isEdit, on: setEdit, off: unsetEdit } = useBoolean(false)
